Reorganise code for new stage activity
