+++ posthog/batch_exports/debug/debugger.py
@@ -27,7 +27,7 @@ from products.batch_exports.backend.temporal.destinations.s3_batch_export import
 from products.batch_exports.backend.temporal.destinations.snowflake_batch_export import (
     snowflake_default_fields,
 )
-from products.batch_exports.backend.temporal.pre_export_stage import (
+from products.batch_exports.backend.temporal.pipeline.internal_stage import (
     _get_s3_staging_folder,
 )
 from products.batch_exports.backend.temporal.spmc import (
@@ -449,7 +449,5 @@ class BatchExportsDebugger:
                 column_stats = ColumnDebugStatistics.from_record_batch(record_batch, column_name=column_name)
             else:
                 column_stats += record_batch
-
-        return column_stats
 
         return column_stats
